demo app: cleanup
