+++ app/src/main/java/cz/mzk/tiledimageview/demonstration/intro/IntroViewmodeFragment.java
@@ -11,7 +11,6 @@ import android.widget.Button;
 
 import cz.mzk.tiledimageview.demonstration.R;
 import cz.mzk.tiledimageview.demonstration.intro.viewMode.IntroViewmodeActivity;
-import cz.mzk.tiledimageview.demonstration.intro.viewMode.IntroViewmodeActivity2;
 
 
 public class IntroViewmodeFragment extends Fragment implements View.OnClickListener {
@@ -32,7 +31,7 @@ public class IntroViewmodeFragment extends Fragment implements View.OnClickListe
         if (mBtnTry == v) {
             FragmentActivity activity = getActivity();
             if (activity != null) {
-                startActivity(new Intent(activity, IntroViewmodeActivity2.class));
+                startActivity(new Intent(activity, IntroViewmodeActivity.class));
             }
         }
