+++ app/src/main/res/layout/activity_intro_viewmode.xml
@@ -1,118 +1,39 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
+<RelativeLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/main_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@color/grey"
-    android:orientation="vertical"
-    android:padding="8dp"
-    >
+    tools:context=".MainActivity">
 
-    <FrameLayout
+    <android.support.v7.widget.Toolbar
+        android:id="@+id/toolbar"
         android:layout_width="match_parent"
-        android:layout_height="0px"
-        android:layout_weight="9"
-        >
+        android:layout_height="wrap_content"
+        android:layout_alignParentTop="true"
+        android:background="?attr/colorPrimary"
+        android:elevation="6dp"
+        android:minHeight="?attr/actionBarSize"
+        android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar"
+        android:title="@string/intro_viewmode_title"
+        app:popupTheme="@style/ThemeOverlay.AppCompat.Light"/>
 
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:background="@color/theme_primary"
-            android:elevation="4dp"
-            android:orientation="vertical"
-            >
-
-            <Spinner
-                android:id="@+id/viewModeSpinner1"
-                android:layout_width="match_parent"
-                android:layout_height="0px"
-                android:layout_alignParentRight="true"
-                android:layout_weight="2"
-                android:padding="4dp"
-                android:popupBackground="@color/white"
-                />
-
-            <cz.mzk.tiledimageview.TiledImageView
-                android:id="@+id/imageView1"
-                android:layout_width="match_parent"
-                android:layout_height="0px"
-                android:layout_margin="16dp"
-                android:layout_weight="6"
-                android:background="@color/theme_primary_light"
-                />
-
-        </LinearLayout>
-
-        <LinearLayout
-            android:id="@+id/progressView1"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:background="@color/progressbar_background_filter"
-            android:gravity="center"
-            >
-
-            <ProgressBar
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:indeterminate="true"/>
-        </LinearLayout>
-
-    </FrameLayout>
-
-    <!--separating space-->
-    <View
+    <android.support.design.widget.TabLayout
+        android:id="@+id/tab_layout"
         android:layout_width="match_parent"
-        android:layout_height="0px"
-        android:layout_weight="1"/>
+        android:layout_height="wrap_content"
+        android:layout_below="@+id/toolbar"
+        android:background="?attr/colorPrimary"
+        android:elevation="6dp"
+        android:minHeight="?attr/actionBarSize"
+        android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar"/>
 
-    <FrameLayout
+    <cz.mzk.tiledimageview.demonstration.intro.viewMode.CustomViewPager
+        android:id="@+id/pager"
         android:layout_width="match_parent"
-        android:layout_height="0px"
-        android:layout_weight="9"
-        >
+        android:layout_height="fill_parent"
+        android:layout_below="@id/tab_layout"/>
 
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:background="@color/theme_primary"
-            android:elevation="4dp"
-            android:orientation="vertical"
-            >
-
-            <Spinner
-                android:id="@+id/viewModeSpinner2"
-                android:layout_width="match_parent"
-                android:layout_height="0px"
-                android:layout_alignParentRight="true"
-                android:layout_weight="2"
-                android:padding="4dp"
-                android:popupBackground="@color/white"
-                />
-
-            <cz.mzk.tiledimageview.TiledImageView
-                android:id="@+id/imageView2"
-                android:layout_width="match_parent"
-                android:layout_height="0px"
-                android:layout_margin="16dp"
-                android:layout_weight="6"
-                android:background="@color/theme_primary_light"
-                />
-        </LinearLayout>
-
-        <LinearLayout
-            android:id="@+id/progressView2"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:background="@color/progressbar_background_filter"
-            android:gravity="center"
-            >
-
-            <ProgressBar
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:indeterminate="true"/>
-        </LinearLayout>
-
-    </FrameLayout>
-
-</LinearLayout>
+</RelativeLayout>
